+++ src/app/modules/slorm-reaper/slorm-reaper-routing.module.ts
@@ -1,0 +1,15 @@
+import { NgModule } from '@angular/core';
+import { RouterModule, Routes } from '@angular/router';
+
+import { ListComponent } from './components/list/list.component';
+
+const routes: Routes = [
+    { path: 'list', component: ListComponent },
+    { path: '**', pathMatch: 'full', redirectTo: 'list' }
+];
+
+@NgModule({
+  imports: [RouterModule.forChild(routes)],
+  exports: [RouterModule]
+})
+export class SlormReaperRoutingModule { }