+++ src/app/modules/slormancer/constants/common.ts
@@ -47,4 +47,5 @@ export const APP_TO_GAME_VERSION_MAPPER: { [key: string]: string } = {
     '0.1.1': '0.3.1071',
     '0.1.2': '0.4.191',
     '0.1.3': '0.4.191',
+    '0.1.4': '0.4.191',
 }