+++ src/environments/environment.ts
@@ -4,7 +4,7 @@
 
 export const environment = {
   production: false,
-  version: '0.1.3',
+  version: '0.1.4',
   githublink: 'https://github.com/cayrac/slorm-planner',
   gameLink: 'http://www.slormitestudios.com'
 };